A plugin dictionary class to solve "dependency" problems.
A nice feature of plugins is that they can be loaded at any time during
the execution of the program. If all plugins must be declared and loaded
at the beginning in order to avoid initialization problem, part of the
fun is removed.

PluginDict is a dictionary to solve this issue. Because it derives from
dict, it is backwards compatible. However, the get and set item are
changed for flexibility. When the program request an item that is not in
the dict, instead of None it returns a placeholder item that can
remember function calls. When that item is set in the future, the calls
are executed.

Many of the plugin handling code in base xmpp could be moved to
PluginDict including convenient functions to load many plugins.

Depedency declaration can be implemented in the plugins and handled
directly at PluginDict. However, I would like to distinguish between two
type of "dependencies"
1.- Plugin A requires Plugin B. This is a real dependency. If A is
loaded, B should be loaded as well.
2.- Plugin A uses Plugin B. Not a dependency. If A is loaded and B is
loaded (now or in the future), A can use B.

I think we need to distinguish between these two cases. But the
PluginDict/CallRecorder/HungryObject is an effective plataform for both.